refactor: remove unused mysql relation
The charm has a mysql relation that is not used as it is provided by the mariadb charm which is no longer supported.
The other alternative is to use the relational-db relation to relate to mysql-k8s charm, but due to #64, it is not possible.

In #71 it was decided that instead of an external DB provider, the mlmd charm will use the SQLite implementation it has been using for a while.
In PR#72 the charm code has been modified to exclusively use SQLite.

Fixes #73