fix: add chown_component to ensure mounted storage belongs to _daemon_ (#119)
* fix: add chown_component to ensure mounted storage belongs to _daemon_

Due to juju/juju#19020, this charm must be responsible for setting the
permissions to the mounted storage requested by the charm.
This charm needs rwx permissions on /data (the path where the storage is mounted)
for it to work properly, and because 16638db changed the pebble service
to be ran as non-root, the permissions are no longer guaranteed;
therefore, this commit adds a chwon_component to the components graph
that directly changes the ownership of the /data directory to the
user defined in the Pebble layer.

Fixes #117